optimize runtime, avoid downloading data/model
