--- tests/trainer/cpu_trainer.py
@@ -1,19 +1,0 @@
-from deepspeed.ops.adam import DeepSpeedCPUAdam
-
-from arctic_training import register
-from arctic_training.optimizer import FusedAdamOptimizerFactory
-
-
-@register
-class CPUAdamOptimizerFactory(FusedAdamOptimizerFactory):
-    name = "cpu-adam"
-
-    def create_optimizer(self, model, optimizer_config):
-        optimizer_grouped_params = self.get_optimizer_grouped_params(
-            model, optimizer_config.weight_decay
-        )
-        return DeepSpeedCPUAdam(
-            optimizer_grouped_params,
-            lr=optimizer_config.learning_rate,
-            betas=optimizer_config.betas,
-        )
